ui-mockups-mvp/requirements.txt
streamlit==1.38.0
pydantic>=2.0.0
python-docx>=1.1.0
pypdf2>=3.0.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
